+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/event/processor/provisioning/ProvisioningUpdateProcessor.java
@@ -57,6 +57,7 @@ public class ProvisioningUpdateProcessor extends AbstractProvisioningProcessor {
 					connectorObject.getObjectClass(), uidAttribute, connectorObject.getAttributes());
 		} else {
 			// TODO: appropriate message - provisioning is not executed - attributes don't change
+			// Operation was logged only. Provisioning was not executes, because attributes does'nt change.
 		}
 		return null;
 	}