+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/service/impl/AbstractProvisioningExecutor.java
@@ -5,6 +5,7 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -61,8 +62,10 @@ import eu.bcvsolutions.idm.acc.service.api.SysSystemAttributeMappingService;
 import eu.bcvsolutions.idm.acc.service.api.SysSystemEntityService;
 import eu.bcvsolutions.idm.acc.service.api.SysSystemMappingService;
 import eu.bcvsolutions.idm.acc.service.api.SysSystemService;
+import eu.bcvsolutions.idm.core.api.domain.CoreResultCode;
 import eu.bcvsolutions.idm.core.api.domain.OperationState;
 import eu.bcvsolutions.idm.core.api.dto.AbstractDto;
+import eu.bcvsolutions.idm.core.api.dto.DefaultResultModel;
 import eu.bcvsolutions.idm.core.api.dto.IdmRoleDto;
 import eu.bcvsolutions.idm.core.api.dto.PasswordChangeDto;
 import eu.bcvsolutions.idm.core.api.entity.OperationResult;
@@ -313,10 +316,7 @@ public abstract class AbstractProvisioningExecutor<DTO extends AbstractDto>
 				accounts.add(entityAccount.getAccount());
 			}
 		});
-
-		// Clear accounts in DTO ... we will set only success changed
-		passwordChange.setAccounts(null);
-		Map<UUID, UUID> operationAccounts = new HashMap<>();
+		Map<UUID, AccAccountDto> operationAccounts = new HashMap<>(); // operationId / account
 		accounts.forEach(accountId -> {
 			AccAccountDto account = accountService.get(accountId);
 			// find uid from system entity or from account
@@ -348,7 +348,7 @@ public abstract class AbstractProvisioningExecutor<DTO extends AbstractDto>
 			SysProvisioningOperationDto operation = prepareProvisioningForAttribute(systemEntity, mappedAttribute, passwordChange.getNewPassword(),
 					ProvisioningOperationType.UPDATE, dto);
 			preparedOperations.add(operation);
-			operationAccounts.put(operation.getId(), account.getId());
+			operationAccounts.put(operation.getId(), account);
 		});
 		passwordChange.setNewPassword(null);
 		passwordChange.setOldPassword(null);
@@ -357,12 +357,28 @@ public abstract class AbstractProvisioningExecutor<DTO extends AbstractDto>
 		return preparedOperations
 			.stream()
 			.map(operation -> {
-				OperationResult result = provisioningExecutor.executeSync(operation).getResult();
-				if (result.getState() == OperationState.EXECUTED) {
+				SysProvisioningOperationDto result = provisioningExecutor.executeSync(operation);
+				Map<String, Object> parameters = new LinkedHashMap<String, Object>();
+				AccAccountDto account = operationAccounts.get(operation.getId());
+				SysSystemDto system = DtoUtils.getEmbedded(account, AccAccount_.system, SysSystemDto.class);
+				//
+				parameters.put("account", ImmutableMap.of(
+						"id", account.getId().toString(),
+						"uid", account.getUid(),
+						"realUid", account.getRealUid(),
+						"systemId", system.getId(),
+						"systemName", system.getName()));
+				if (result.getResult().getState() == OperationState.EXECUTED) {
 					// Add success changed password account
-					passwordChange.getAccounts().add(operationAccounts.get(operation.getId()).toString());
+					return new OperationResult
+							.Builder(OperationState.EXECUTED)
+							.setModel(new DefaultResultModel(CoreResultCode.PASSWORD_CHANGE_ACCOUNT_SUCCESS, parameters))
+							.build();
 				}
-				return result;
+				return new OperationResult
+						.Builder(result.getResult().getState())
+						.setModel(new DefaultResultModel(CoreResultCode.PASSWORD_CHANGE_ACCOUNT_FAILED, parameters))
+						.build();
 			})
 			.collect(Collectors.toList());
 	}
@@ -510,6 +526,7 @@ public abstract class AbstractProvisioningExecutor<DTO extends AbstractDto>
 			List<? extends AttributeMapping> attributes) {
 		AccAccountDto account = getAccountSystemEntity(systemEntity.getId());
 		String uid = systemEntity.getUid();
+		SysSystemDto system = DtoUtils.getEmbedded(systemEntity, SysSystemEntity_.system, SysSystemDto.class);
 		Map<ProvisioningAttributeDto, Object> accountAttributes = new HashMap<>();
 		 
 		// delete - account attributes is not needed
@@ -530,7 +547,7 @@ public abstract class AbstractProvisioningExecutor<DTO extends AbstractDto>
 				Object uidValue = getAttributeValue(uid, dto, attribute);
 				if (!(uidValue instanceof String)) {
 					throw new ProvisioningException(AccResultCode.PROVISIONING_ATTRIBUTE_UID_IS_NOT_STRING,
-							ImmutableMap.of("uid", uidValue));
+							ImmutableMap.of("uid", uidValue, "system", system.getName()));
 				}
 				updateAccountUid(account, uid, (String)uidValue);
 				accountAttributes.put(ProvisioningAttributeDto.createProvisioningAttributeKey(attribute, schemaAttributeDto.getName()), uidValue);
@@ -554,7 +571,7 @@ public abstract class AbstractProvisioningExecutor<DTO extends AbstractDto>
 					.createProvisioningAttributeKey(attributeParent, schemaAttributeParent.getName());
 			if (!schemaAttributeParent.isMultivalued()) {
 				throw new ProvisioningException(AccResultCode.PROVISIONING_MERGE_ATTRIBUTE_IS_NOT_MULTIVALUE,
-						ImmutableMap.of("object", uid, "attribute", schemaAttributeParent.getName()));
+						ImmutableMap.of("object", uid, "attribute", schemaAttributeParent.getName(), "system", system.getName()));
 			}
 
 			List<Object> mergedValues = new ArrayList<>();
@@ -592,9 +609,9 @@ public abstract class AbstractProvisioningExecutor<DTO extends AbstractDto>
 	}
 	
 	@Override
-	public OperationResult doProvisioningForAttribute(SysSystemEntityDto systemEntity, AttributeMapping attributeMapping,
+	public void doProvisioningForAttribute(SysSystemEntityDto systemEntity, AttributeMapping attributeMapping,
 			Object value, ProvisioningOperationType operationType, DTO dto) {
-		return provisioningExecutor.execute(prepareProvisioningForAttribute(systemEntity, attributeMapping, value, operationType, dto)).getResult();
+		provisioningExecutor.execute(prepareProvisioningForAttribute(systemEntity, attributeMapping, value, operationType, dto));
 	}