+++ CHANGELOG.md
@@ -114,6 +114,7 @@ All notable changes to this project will be documented in this file.
 ##### Identity
 
 - Deprecated method ``IdmIdentityService#saveIdentity`` was removed. Use ``save``or ``publish`` method instead.
+- Method ``IdmIdentityService#passwordChange`` don't modify given password change dto - results was added.  
 
 ##### Event processiong
 
@@ -134,8 +135,7 @@ All notable changes to this project will be documented in this file.
 
 ##### Notifications
 
-- Api [NotificationSender](https://github.com/bcvsolutions/CzechIdMng/blob/develop/Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/notification/api/service/NotificationSender.java) was changed. Now are returned all notifications that is sent. Also mechanism of sending with topic was fixed. Check your project notification configuration.
-- Method resolveTemplate in [IdmNotificationTemplateService](https://github.com/bcvsolutions/CzechIdMng/blob/develop/Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/notification/api/service/IdmNotificationTemplateService.java) was changed. Now this method return only found template, wildcard is not returned anymore.
+- Api ``NotificationSender`` was changed. Now are returned all notifications that is sent. Also mechanism of sending with topic was fixed. Check your project notification configuration.
 
 
 ##### Role
@@ -279,7 +279,9 @@ All notable changes to this project will be documented in this file.
 - Entity **SysProvisioningArchive was transformed to SysProvisioningArchiveDto**. Update all places where call from your project [SysProvisioningArchiveService](https://github.com/bcvsolutions/CzechIdMng/blob/develop/Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/service/api/SysProvisioningArchiveService.java). The service now uses only DTO.
 - Entity **SysProvisioningOperation was transformed to SysProvisioningOperationDto**. Update all places where call from your project [SysProvisioningOperationService](https://github.com/bcvsolutions/CzechIdMng/blob/develop/Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/service/api/SysProvisioningOperationService.java). The service now uses only DTO.
 - Entity **SysProvisioningRequest was removed** - use ``SysProvisioningOperation`` and ``SysProvisioningOperationDto``.
-- Method **SysProvisioningBatch findBatch(SysProvisioningOperation operation)** from service [SysProvisioningBatchService](https://github.com/bcvsolutions/CzechIdMng/blob/develop/Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/service/api/SysProvisioningBatchService.java) was moved to service [SysProvisioningOperationService](https://github.com/bcvsolutions/CzechIdMng/blob/develop/Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/service/api/SysProvisioningOperationService.java), update all places where you call method **findBatch** to new service.
+- Method **SysProvisioningBatch findBatch(SysProvisioningOperation operation)** from service ``SysProvisioningBatchService`` was moved into service ``SysProvisioningOperationService``, update all places where you call method **findBatch** to new service.
+- Password provisioning was improved - operation results are returned now, without input password dto is changed.
+- Provisioning operation table is sorted by created date descended.
 
 ### Removed